+++ packages/moonstone/CHANGELOG.md
@@ -14,7 +14,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
--`moonstone/IconButton` Tooltip position when disabled
+- `moonstone/IconButton` Tooltip position when disabled
+- `moonstone/VideoPlayer` to display feedback tooltip correctly when navigating in 5-way
 
 ## [1.7.0] - 2017-08-23